[2216] Fix a possible crash related to Broggok event

+++ scripts/outland/hellfire_citadel/blood_furnace/blood_furnace.h
@@ -82,7 +82,7 @@ class MANGOS_DLL_DECL instance_blood_furnace : public ScriptedInstance
         BroggokEventInfo m_aBroggokEvent[MAX_ORC_WAVES];
 
         uint32 m_uiBroggokEventTimer;                       // Timer for opening the event cages; only on heroic mode = 30 secs
-        uint8 m_uiBroggokEventPhase;
+        uint32 m_uiBroggokEventPhase;
 
         GUIDList m_luiNascentOrcGUIDs;
 };
